Modified permissions. Various styling tweaks. Integrated adaptive backgrounds. Added custom widgets.
Modified permissions of the theme files. Changed the owner to www-data:www-data and then set the perms to 755 as per the WordPress spec.

Various styling tweaks. Moved the post meta to the top right. Changed the read more button a bit. Began styling the footer.

Integrated adaptive backgrounds for the featured images of posts. On load the jquery.adaptive-backgrounds.js script will process the featured image, find the dominant color, and then apply that as the background color for the parent div.section.

Added custom text widgets for the about me text and gravatar. Added two widget areas for the footer, one upper and one lower.