Removed Admin Bar. Moved footer to custom function. Footer style fixes. Single fixes.
Removed the admin bar as the margins were messing up with the full page layout and things were getting cut off.

Moved the footer to a custom function so it can be called from multiple parts of the theme. Fixed styling in the footer.

Fixed issues with the single where the page was getting rendered as a full page. Also fixed the handling of the footer so that the shadow overflow is fixed.