Fixed major issue with anchors. Changed anchors to post slug. Added fullPage.js stylesheet.
Fixed major issue with anchors where the slides would skip around when scrolling. This was happening because the ID's on the div.section's were being set to the same as the anchors were. This was fixed by adding the data-anchor attribute to each div.section. jQuery then grabs all of these anchor datas and throws them into an array for fullPage to use.

Added the fullPasge.js stylesheet on load to make sure we're getting all the styles we need.