Fixed scroll being broken on single pages
Mouse wheel scrolling would cease to function due to fullPage.js's stylesheet that applies overflow: hidden to both the html and body elements. A fix was put into place that adds the class wp-single to the html element if the current page is a single page/post. The html element is then selected in CSS with the NOT operand so it will not select html.wp-single.